fix: Addressed Dakota's revisions on Login and Signup screens.
The keyboard does not block the textfil
eds anymore on both Login and Signup. On sign up, I added two textfields, one for firstName and another for lastName on the Signup screen. Also followed Dakota's suggestion to use navigation.replace('Login') when the SignUp button is pressed. And lastly, the "Dont have an account\? Sign Up" message is properly aligned.